ops-platform/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
prometheus-api-client==0.4.0
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
